list and vector list sort uisng circular nodes Signed-off-by: min.song@arcserve.com <min.song@arcserve.com>

+++ TaskRunner/SmlVectorList.h
@@ -300,7 +300,7 @@ namespace SmartLib
 			
 
 			long minId = _INVALID_ID;
-			for (int ii = 0; ii < ARRAY_SIZE; ++ii)
+			for (long ii = 0; ii < ARRAY_SIZE; ++ii)
 			{
 				if (arr[ii] >= 0)
 				{
